minor fix, stub out good pages
