+++ webapp/app/routes/good.building.tsx
@@ -1,0 +1,9 @@
+import { Layout } from "../components/Layout";
+
+export default function GoodBuilding() {
+  return (
+    <Layout>
+      <h1>Good Building</h1>
+    </Layout>
+  );
+}
